Fix de eliminar, en caso de refrescar no reintente eliminar

+++ controllers/default.py
@@ -197,12 +197,14 @@ def registrousrth_final():
 
 def eliminarusrth():
     if request.args:
+        
         id_persona = int(request.args[0])
-        print id_persona
-        bombero = db(db.bombero.id_persona==id_persona).select(db.bombero.id_usuario).first()
-        id_usuario = bombero.id_usuario
-        db(db.persona.id==id_persona).delete()
-        db(db.usuario.id==id_usuario).delete()
+        
+        if not db(db.persona.id==id_persona).isempty():
+            bombero = db(db.bombero.id_persona==id_persona).select(db.bombero.id_usuario).first()
+            id_usuario = bombero.id_usuario
+            db(db.persona.id==id_persona).delete()
+            db(db.usuario.id==id_usuario).delete()
 
     tabla = db(db.persona).select(join=db.bombero.on(db.bombero.id_persona == db.persona.id))
